improvement: response functions now simpler to use for most cases
Response functions no longer have to be async. For the majority of cases the response does not need
to be nested on an object with a response property. More advanced cases now need to use __override
when responseCode, responseHeaders or delay needs to be customised within the same response
function.

BREAKING CHANGE: The return object for response functions do not use a nested object with a response
property any more. The solution is to either remove the nested response or for more advanced use
cases to nest under a further __override property.