feat: extended updateContext so that it can also accept a function
The updateContext function passed to response handlers used to only accept a partial context object,
which merges into the context object. It now also accepts a function. This function will be passed
the current context and should return a partial object which will be merged into the context object.
This is to enable async processes like setTimeout to access the most up to date context.